bugfix0.094

+++ ais-bench_workload/tool/find_best_batchsize/test/test_aoe_mode.sh
@@ -6,6 +6,7 @@
 # --input_shape_str 需要依据具体模型配置
 
 declare -i ret_ok=0
+declare -i ret_run_failed=1
 
 # 功能测试 
 echo "test: model_type onnx, aoe_mode on, max_batch_num = 64"
@@ -21,7 +22,6 @@ cmd="bash ../find_best_batchsize.sh
     --job_type 1
     "
 [ $cmd != $ret_ok ] && { echo "command failed";return 1; }
-#$cmd
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -37,8 +37,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 0
     --job_type 1
     "
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[ $cmd != $ret_ok ] && { echo "command failed";return 1; }
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -55,8 +54,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 1
     --job_type 1
     "
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[ $cmd != $ret_ok ] && { echo "command failed";return 1; }
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -74,8 +72,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 1
     --job_type 1
     "
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[ $cmd != $ret_ok ] && { echo "command failed";return 1; }
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -93,7 +90,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 3
     --job_type 1
     "
-$cmd
+[ $cmd != $ret_run_failed ] && { echo "not recognize illegal command";return 1; }
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -110,6 +107,6 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 1
     --job_type 3
     "
-$cmd
+[ $cmd != $ret_run_failed ] && { echo "not recognize illegal command";return 1; }
 cmd="rm -rf ../cache/"
 $cmd
